+++ src/cli/import.rs
@@ -22,6 +22,10 @@ pub struct ImportCommand {
 #[derive(Copy, Clone, Debug, ValueEnum)]
 enum Format {
     Insomnia,
+    /// A Jetbrains `.http` file in the REST format (without Jetbrains env files)
+    Jetbrains,
+    /// A VSCode `.rest` file in the REST format
+    Vscode,
 }
 
 impl Subcommand for ImportCommand {
@@ -29,6 +33,8 @@ impl Subcommand for ImportCommand {
         // Load the input
         let collection = match self.format {
             Format::Insomnia => Collection::from_insomnia(&self.input_file)?,
+            Format::Vscode => Collection::from_vscode(&self.input_file)?,
+            Format::Jetbrains => Collection::from_jetbrains(&self.input_file)?,
         };
 
         // Write the output